+++ db/migrate/20260923160000_change_notify_mode_default_to_users.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ChangeNotifyModeDefaultToUsers < ActiveRecord::Migration[7.1]
+  def up
+    change_column_default :mattermost_project_settings, :notify_mode, "users"
+  end
+
+  def down
+    change_column_default :mattermost_project_settings, :notify_mode, "group"
+  end
+end